+++ CHANGELOG.md
@@ -52,3 +52,7 @@ Corrected a minor error in the README documentation.
 
 ## Version 1.1.4 - 2025/12/27 - ITA
 Added the outstanding change log for the previous version.
+
+## Version 1.1.5 - 2025/12/28 - ITA
+- Improved the documentation of functions to show better on the tooltip in IDEs.
+- Improved the deepClone function to handle Date and array types that may have dates types better.